Tweaked plotimage.php to optimize the logic.  When drawing the tic marks, only need to set $plot_y once per iteration of the outer for($y....) loop.  And again for the X tic marks loop.
Added comments to plotimage.php.

Following are comments for my previous check-in, but I forgot to add it in time.
Added uploadfile.php which contains a copy of the bobdemo class
and has conditional logic to insert ed uploaded file into MySQL if Server Method is 'POST',
and returns an HTML form to provide an <input type="file"/> form to post to uploadfile.php